Add files via upload
This file is written in the Java programming language. The program will ask the user to input two integers and will have an option to choose from a list of whether they want to add, subtract, multiply or divide as well as an option to quit the program. Whatever option is chosen it will output the result of the two numbers entered.